Move _string_dict_from_arrow_schem to conversions.py
