+++ apis/python/tests/test_registration_mappings.py
@@ -17,7 +17,6 @@ import pytest
 import tiledbsoma.io
 import tiledbsoma.io._registration as registration
 from tiledbsoma.io import conversions
-from tiledbsoma.io._registration import signatures
 
 from ._util import assert_adata_equal
 
@@ -256,7 +255,7 @@ def test_pandas_indexing(
             df.index.name = index_col_and_name[1]
 
     arrow_schema = conversions.df_to_arrow_schema(df, default_index_name)
-    actual_signature = signatures._string_dict_from_arrow_schema(arrow_schema)
+    actual_signature = conversions._string_dict_from_arrow_schema(arrow_schema)
     expected_signature = dict((col, "string") for col in signature_col_names)
     assert actual_signature == expected_signature
 # fmt: on
